requirements.txt
beautifulsoup4==4.12.3
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
packaging==26.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
